src/requirements.txt
pyyaml
requests

boto3 >= 1.17.12